+++ src/main/java/com/alexnmat/exam/models/DTO/SubTaskDTO.java
@@ -9,8 +9,8 @@ public class SubTaskDTO {
     private LocalDate utilStartDate;
     private LocalDate utilEndDate;
     private Boolean completed;
-    private long allocatedHours;
-    private int totalTimeSpent;
+    private double allocatedHours;
+    private double totalTimeSpent;
 
     public SubTaskDTO(long id, String subTaskName, LocalDate utilStartDate, LocalDate utilEndDate, Boolean completed) {
         this.id = id;
@@ -20,25 +20,25 @@ public class SubTaskDTO {
         this.completed = completed;
     }
 
-    public SubTaskDTO(long id, long allocatedHours, int totalTimeSpent) {
+    public SubTaskDTO(long id, double allocatedHours, double totalTimeSpent) {
         this.id = id;
         this.allocatedHours = allocatedHours;
         this.totalTimeSpent = totalTimeSpent;
     }
 
-    public long getAllocatedHours() {
+    public double getAllocatedHours() {
         return allocatedHours;
     }
 
-    public void setAllocatedHours(long allocatedHours) {
+    public void setAllocatedHours(double allocatedHours) {
         this.allocatedHours = allocatedHours;
     }
 
-    public int getTotalTimeSpent() {
+    public double getTotalTimeSpent() {
         return totalTimeSpent;
     }
 
-    public void setTotalTimeSpent(int totalTimeSpent) {
+    public void setTotalTimeSpent(double totalTimeSpent) {
         this.totalTimeSpent = totalTimeSpent;
     }
