+++ src/main/java/com/alexnmat/exam/models/entities/SubProject.java
@@ -45,10 +45,10 @@ public class SubProject implements Serializable {
     private String description;
 
     @Column(name = "allocated_hours")
-    private long allocatedHours;
+    private double allocatedHours;
 
     @Column(name = "total_time_spent")
-    private int totalTimeSpent;
+    private double totalTimeSpent;
 
     @OneToMany(mappedBy = "subProject", cascade = CascadeType.MERGE)
     private List<Task> tasks;
@@ -120,19 +120,19 @@ public class SubProject implements Serializable {
         this.description = description;
     }
 
-    public long getAllocatedHours() {
+    public double getAllocatedHours() {
         return allocatedHours;
     }
 
-    public void setAllocatedHours(long allocatedHours) {
+    public void setAllocatedHours(double allocatedHours) {
         this.allocatedHours = allocatedHours;
     }
 
-    public int getTotalTimeSpent() {
+    public double getTotalTimeSpent() {
         return totalTimeSpent;
     }
 
-    public void setTotalTimeSpent(int totalTimeSpent) {
+    public void setTotalTimeSpent(double totalTimeSpent) {
         this.totalTimeSpent = totalTimeSpent;
     }
