+++ src/main/java/com/alexnmat/exam/models/entities/SubTask.java
@@ -48,10 +48,10 @@ public class SubTask implements Serializable {
     private String description;
 
     @Column(name = "allocated_hours")
-    private long allocatedHours;
+    private double allocatedHours;
 
     @Column(name = "total_time_spent")
-    private int totalTimeSpent;
+    private double totalTimeSpent;
 
     public SubTask() {
     }
@@ -112,19 +112,19 @@ public class SubTask implements Serializable {
         this.description = description;
     }
 
-    public int getTotalTimeSpent() {
+    public double getTotalTimeSpent() {
         return totalTimeSpent;
     }
 
-    public void setTotalTimeSpent(int totalTimeSpent) {
+    public void setTotalTimeSpent(double totalTimeSpent) {
         this.totalTimeSpent = totalTimeSpent;
     }
 
-    public long getAllocatedHours() {
+    public double getAllocatedHours() {
         return allocatedHours;
     }
 
-    public void setAllocatedHours(long allocatedHours) {
+    public void setAllocatedHours(double allocatedHours) {
         this.allocatedHours = allocatedHours;
     }
 }
